separate join_room logic with whole game start. Join room currently works as it should and frontend fetching username & avatar
