+++ frontend/static/css/homepage.css
@@ -10,7 +10,7 @@ body {
   display: flex;
   flex-direction: column;
   align-items: center;
-  overflow-y: hidden;
+  /*overflow-y: hidden;*/
 }
 
 /* Title */
